// src/Polyglot.Frontend/src/app/api/model/availableModelDto.ts
/**
 * Polyglot.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AvailableModelDto { 
    id?: string | null;
    name?: string | null;
    contextLength?: number;
    inputModalities?: Array<string> | null;
    outputModalities?: Array<string> | null;
    promptPricePerMillion?: number;
    completionPricePerMillion?: number;
}