// src/Polyglot.Frontend/src/app/api/model/chatStreamPayloadType.ts
/**
 * Polyglot.API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export const ChatStreamPayloadType = {
  Chunk: 'Chunk',
  Reasoning: 'Reasoning',
  Done: 'Done',
  Error: 'Error',
  ToolCall: 'ToolCall',
  ToolResult: 'ToolResult',
} as const;
export type ChatStreamPayloadType =
  (typeof ChatStreamPayloadType)[keyof typeof ChatStreamPayloadType];