// src/Polyglot.Frontend/src/app/api/model/messageDto.ts
/**
 * Polyglot.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MessageRole } from './messageRole';


export interface MessageDto { 
    id?: string;
    role?: MessageRole;
    content?: string | null;
    model?: string | null;
    toolCalls?: string | null;
    toolCallId?: string | null;
    finishReason?: string | null;
    sequenceNumber?: number;
    createdAt?: string;
}
export namespace MessageDto {
}