// src/Polyglot.Frontend/src/app/api/model/sendMessageDto.ts
/**
 * Polyglot.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MessageDto } from './messageDto';


export interface SendMessageDto { 
    chatId?: string;
    userMessage?: MessageDto;
    assistantMessage?: MessageDto;
}