// src/Polyglot.Frontend/src/app/api/model/messageRole.ts
/**
 * Polyglot.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const MessageRole = {
    User: 'User',
    Assistant: 'Assistant',
    System: 'System',
    Tool: 'Tool'
} as const;
export type MessageRole = typeof MessageRole[keyof typeof MessageRole];